Removing snapshot of old hpcc versions
[approved by sungeun, hilde]

We snapshotted this early version of the HPCC benchmarks out of...
sentimentality?  a sense that they might test things that more
recent versions wouldn't?  I'm not sure, but I'm no longer convinced
that keeping them around is worth anything.  SVN will keep the
history for us, and the fewer the number of variations of these
that we keep alive, the less redundancy there will be in trying
to keep them all working and moving forward.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20836 3a8e244f-b0f2-452b-bcba-4c88e055c3ca