Removing limit of stacksize in start_test
[agreed to in principle by Greg]

On the latest cygwin install on our laptops, this 'limit stacksize'
line is causing portability problems that break testing.  The limit
dates back to r3015 and seems like it may have been a little bit
superstitious/mysterious on my part.  In any case, it's hard to
believe we rely on it, so we'll turn it off and see if anything
breaks badly.  Arguably, if a specific test relies on this, we
should restrict the limit specifically for that test rather than
doing a blanket limit.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20837 3a8e244f-b0f2-452b-bcba-4c88e055c3ca