Restructure internal modules corresponding to particular runtime layers.
Currently the internal modules that correspond to particular runtime layers
are directly in modules/internal, for example modules/internal/ugni/*.
We're going to move these in order to match the directory hierarchy used
elsewhere.  Specifically, we're going to interpose a comm, tasks, or threads
component where appropriate.  So for example, modules/internal/ugni will
become modules/internal/comm/ugni.

My overall plan for this change is to create the new directories, change the
compiler to accept both, wait a while, and then at some later time (in a
month or two) remove the old directories after warning everyone that they
must have an updated compiler.

This is the first step in this process.  It creates comm, tasks, and threads
directories in modules/internal.  It also creates modules/internal/comm/ugni
and moves modules/internal/ugni/NetworkAtomicTypes.chpl to that new dir.
Finally, in place of modules/internal/ugni/NetworkAtomicTypes.chpl it adds a
symbolic link of the same name, which refers to the moved copy of the file.
Thus the actual file contents are only in the new place, but they still also
appear to be present in the old place.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20838 3a8e244f-b0f2-452b-bcba-4c88e055c3ca