Factor the description of sync and single variables out of ChapelBase, in preparation for merging locale-aware allocation from the caseyb branch.
This factoring is necessary, because the implementation of sync variables
involves allocation (since they are class instances).  Locale-aware allocation in turn
depends on the default locale implementation (currently in ChapelLocale).  Thus,
the _syncvar and _singlevar classes must be declared after locales are.

The simple solution is to factor syncs and singles out of ChapelBase, and then
mention the new module in ChapelStandard.chpl, somewhere before its first use in
ChapelArray.chpl.

Reviewer: bradc


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20852 3a8e244f-b0f2-452b-bcba-4c88e055c3ca