Change remaining instances of "x[te]-cle" platform name to "cray-x[te]".
[Reviewed by Sung.]

I'm about to remove support for the old xt-cle and xe-cle platform
names, so I made another pass through the tree and found a few last
instances of the old names.  I'm changing those to the corresponding
cray-xt or cray-xe names here.

Note that in the case of the nightly.chpbld0? script files, changing
the platform name won't change the log filenames, because those are
based on the output of the util/chplenv/platform utility, and that
was already doing the translation from the old platform name to the
new one.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20854 3a8e244f-b0f2-452b-bcba-4c88e055c3ca