In the flatten functions pass, pass more variables by value.
[Reviewed by Sung]

The following variables are now passed by value:

- distribution/domain/array wrappers;
- constants and coforall index variables, if their type is one of:
    bool, int, real, imag, complex, enum, class, chpl_taskID_t, range.

Previously these were passed by reference in the flatten function pass
then in some cases switched to by-value by the remote value forwarding
optimization.

A special treatment is given to coforall index variables because they
are not marked as constants (i.e. with FLAG_CONST), even though they
actually *are* constant within each task that's created by that
coforall. Likewise, distribution/domain/array wrappers are constant, too.

(To elaborate on the above for the wrappers:
The const-ness check fires upon seeing FLAG_DISTRIBUTION, FLAG_DOMAIN,
FLAG_ARRAY. These indicate that the symbol is one of _distribution,
_domain, _array wrapper records. Those correspond to source-level
dmap/domain/array variables. Each variable (at the Chapel level)
always refers to the same array/etc. throughout its
lifetime. Therefore the pointer that's sitting in the corresponding
_array/etc. record always points to the same descriptor object. It is
that pointer that I consider constant and pass by value. The object
itself, i.e. its fields, can surely change, but the fields are not
passed by value in this commit.)

Passing a constant by value is correct in the sense that the contents
of memory occupied by that constant does not change. It is beneficial
in particular when the function is invoked on a different locale -
which is the case with the functions created to implement 'on' clauses.

A couple of comm diagnostic tests have been updated to reflect the improvement.

There is room for improvement. We can pass more things by value,
e.g. records and tuples (perhaps only up to a certain size) and object
fields (esp. useful for array descriptors). We can regularize the code
by marking the coforall index variables and distribution/domain/array
wrappers with FLAG_CONST. With this commit, wrapper records and range
records *are* considered for by-value passing using special cases.
Because we know what those records contain/how big they are.
I am postponing a general routine to decide whether to consider
a general user record.

While there, I factored out the check for "should we pass this
variable by reference" into a separate function.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20857 3a8e244f-b0f2-452b-bcba-4c88e055c3ca