Improved locality in edge generation in the release version of SSCA2.
[reviewed by Sung]

r20857 allowed me to be more aggressive in using 'local' blocks,
so I added one and extended the scope of another.

I still have to create a const copy of the 'bit' variable,
because 'bit' itself does vary and the remote value forwarding
is too conservative in this case to pass it by value.

The second 'local' I added, before

  w = floor (1 + rnd * MAX_EDGE_WEIGHT) : int

- should not affect performance. I like it to document my expectations
that the variables inside it do not need communication.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20866 3a8e244f-b0f2-452b-bcba-4c88e055c3ca