Correct INFINITY and NAN type and provide XMT implementation.
[Reviewed by Sung.]

The spec specifies that the predefined INFINITY and NAN values in the
standard math module are of type real(64), but the implementation had
them as real(32).  Fix this.

As a side effect, this also fixes a separate problem on XMT systems.
There, compiling the emitted C code for a Chapel program was failing
because the LLVM commit assumed that all targets were POSIX-based and
had the INFINITY and NAN predefined values in their <math.h>.  The XMT
is BSD-based and does not.  The new code resolves this problem because
POSIX defines INFINITY and NAN to be of type float; since we need
double, we no longer refer to these standard C library values on any
platform.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20873 3a8e244f-b0f2-452b-bcba-4c88e055c3ca