The communication diagnostics and memory tracking interfaces were removed from the ChapelLocale module, in preparation for merging changes supporting hierarchical locales.
Moved the communication diagnostics from modules/internal/ChapelLocale.chpl to a
new file modules/standard/CommDiagnostics.chpl, and updated tests which must now
mention this module explicitly.  Also added mention to BlockDist.chpl, since it
uses that module if diagnostics are enabled.  This may be redundant with the
mention in some of the tests, but that is OK.
    modules/internal/ChapelLocale.chpl
    modules/standard/CommDiagnostics.chpl
    test/optimizations/sungeun/RADOpt/access.chpl
    test/optimizations/rafa/PrintComms.chpl
    test/optimizations/bulkcomm/alberto/PrintComms.chpl
    test/parallel/taskPar/sungeun/barrier/commDiags.chpl
    test/performance/sungeun/assign_across_locales.chpl
    modules/dists/BlockDist.chpl

Moved the memory tracking interface to its own file
modules/internal/MemTracking.chpl, and mentioned it in ChapelStandard.chpl.
    modules/internal/ChapelLocale.chpl
    modules/internal/ChapelStandard.chpl
    modules/internal/MemTracking.chpl

Reviewed by: bradc, sungeun


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20874 3a8e244f-b0f2-452b-bcba-4c88e055c3ca