Remove support for old platform names "xt-cle" and "xe-cle".
Several months ago the platform names xt-cle and xe-cle were changed to
cray-xt and cray-xe, respectively, and the old names were deprecated.
Now that the release is out, it's time to remove support for those old
deprecated names in the development trunk.

(I didn't have this reviewed because it's really small and we'd already
discussed it.)


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20875 3a8e244f-b0f2-452b-bcba-4c88e055c3ca