Fix for synchronization in gasnet chpl_comm_private_broadcast() for large messages
[reviewed by gbt]

This commit fixes a bug in the gasnet implementation of
chpl_comm_private_broadcast().  For large message, we chunk up the
broadcast data and send 2 or more smaller messages.  Each call
requests the gasnet callback (AM_signal()) to indicate when it has
completed by writing into a 'done' location.  We were only waiting for
one of the calls to flag completion.  This led to cases where the
callback function was writing into free'ed memory.

This commit makes the initiating task wait for all the broadcast
messages to be completed.  I did this by making AM_signal() execute an
atomic fetch-add on the 'done' location and waiting for done to equal
the number of messages sent.  At Greg's suggestion, I also changed the
type of all the 'done' variables to explicit int32_t to match the type
signature of the atomic routine.

This should fix the intermittent time outs and failures of
studies/hpcc/RA/testRAStream.chpl.  Note that this bug was tickled by
a broadcast of a 64*int tuple in the initialization of RARandomStream
module.  Valgrind showed that the release RA also exhibited the same
behavior, but for some reason it did not result in failures.  It's
possible that the values being broadcast were not used as quickly in
the full RA or that the free'ed memory was not reused right away.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20877 3a8e244f-b0f2-452b-bcba-4c88e055c3ca