Remove alloca from QIO
QIO was using alloca in some cases (to avoid calling malloc
when we only needed a small pointer). However, this use of
alloca was causing problems with PGI testing and although
alloca is normally supported it isn't actually standard.

This patch changes the relevant functions and macros
to use a fixed-size stack buffer (since the maximum
amount of data we would alloca is pretty small so we
might as well just reserve space for it all on the stack).

We could use C99 variable-length arrays but that would
make the code more complex... and if it turns out
that we're using too much stack space for these buffers
we can reasonably reduce MAX_ON_STACK (which is just 128 bytes).

I've tested this patch on tests/io.

Thanks to Tom and Greg for their helpful reviews.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20879 3a8e244f-b0f2-452b-bcba-4c88e055c3ca