Add a few new memory leak tests for reference counted data types
I've been using these for my ref counting work, and I thought it'd be
good to check in now as is.

NOTE: As we know, we leak memory for non-default distributions, so the
.goods for the domainMaps.chpl test do not report all 0's.  I'm
checking this in so make sure we don't make the problem worse going
forward.  Hopefully, it won't be too noisy.  There is also a leak in
the test case for the default distribution that I think should be
attributed to leaking strings (though the number changes for
CHPL_COMM=gasnet with -nl 1, so maybe not?).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20883 3a8e244f-b0f2-452b-bcba-4c88e055c3ca