Adding 'use CommDiagnostics' to comm diags tests
These were missed in the inital commit of the CommDiagnostics module
(r20874) and subsequent updates.

Note that for the hpcc comm diags tests, I put the use statement in
the file generated by the PRECOMP.  The only test failing in that
directory was stream-ep_commDiags.chpl because all the others use
BlockDist and that uses CommDiagnostics.

In my testing, there appears to be a further reduction in comm counts
for the syncsingle test.  I don't know when this started happening.  I
don't see any notes in the TRIAGE or REGRESSIONS file.  I'll wait to
update until after tonight's testing.  It's possible it was missed by
Vass with his flatten functions changes (r20857).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20899 3a8e244f-b0f2-452b-bcba-4c88e055c3ca