Bit patterns for double INFINITY and NAN must be 64 bits on all systems.
I made a mistake in my r20873 commit -- "unsigned long" isn't 64 bits
everywhere; in particular, on linux32 systems it's only 32 bits.  This
commit changes the bit patterns corresponding to double INFINITY and NAN
to "unsigned long long" and uses the "ULL" suffix on the corresponding
constants.  This will result in 64-bit values on all systems, according
to the C99 standard.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20901 3a8e244f-b0f2-452b-bcba-4c88e055c3ca