Adapt fully to the atomics interface.
[Reviewed by Sung.]

In an earlier commit (r20877) we begin using the atomics interface to
count the sub-messages sent for large private broadcasts.  We didn't
realize it at the time, but that didn't work with CHPL_ATOMICS=locks due
to our assumption that the atomic types were synonyms for their integer
equivalents.  While that's true when CHPL_ATOMICS=intrinsics, in the
locks case the atomic datum is a actually struct, containing a sync var
for protection plus the protected integer value.  As a result, on
linux32 platforms where CHPL_ATOMICS=locks by default due to lack of
hardware support for the full range of atomic ops, comm-gasnet.c didn't
compile.

This commit changes the type of the object that is the target of GASNet
signalling operations so that everything works with either kind of
atomics.  We use the atomics interface to initialize and increment a
counter, and when that hits a preset target value, we flip an integer
flag from 0 to 1 to indicate completion.  The GASNET_BLOCKUNTIL()
operations just wait on the flag.

While here, I also somewhat cleaned up how we build and pass address
parts through GASNet requests to the signalling callbacks.

I also fixed a bug.  In chpl_comm_pre_task_exit() we were directly
storing to a flag to tell the polling task to exit.  But this flag was
the subject of a GASNET_BLOCKUNTIL in the polling task itself, and the
GASNet documentation says that GASNet can assume that the only Active
Message handlers will change values involved in such conditions.  So the
direct store was illegal, and I replaced it with an appropriate signal
call.  (See r15157 for a previous discovery of this same restriction.)

The new code passed the multilocale suite on both lin64 and linux32.

I ran pre-change and post-change RA executables on 128 locales of a Cray
XE system, with n=38 and N_U=1<<24, and saw no significant difference in
performance between the two.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20904 3a8e244f-b0f2-452b-bcba-4c88e055c3ca