Allow --gen-ids and other compiler switches to be negated [Reviewed by sungeun]
The option of using a "no-" prefix was added to the following compiler flags:
     count-tokens     print-code-size     print-search-dirs     baseline
     llvm     print-commands     print-passes

The criteria for selecting these flags were:
 - They are Boolean and false by default.
 - They have a corresponding environment symbol.
 - They do not have a prefix no- already.

userhelp.good was updated to match the new --help output.

chpl.txt was updated to reflect the new option syntax.

The --no-gen-ids option was added to gen_extern_prototypes.compopts.  This
actually drove all of the above changes, since setting CHPL_GEN_IDS in the
environment caused the test to fail.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20907 3a8e244f-b0f2-452b-bcba-4c88e055c3ca