Updating entry on minus_equals test
Updated the entry on domains/sungeun/assoc/minus_equals.chpl to
include notes from last year when I found this race condition.
Basically, we overwrite a shared variable (val iterator index) and
then copy the value out (no locking).  Ideally, we should not be
writing into the shared variable at all, since we just copy the value
out.

We don't run into this more often is because the window for races is
typically very small (usually between 2 lines of the generated code),
but it's slightly larger for --no-local for associative arrays
promoted remove().



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20909 3a8e244f-b0f2-452b-bcba-4c88e055c3ca