Clean up a few bugs in the bytes implementation
A few functions were renamed but not updated in the bytes implementation in
Buffers.chpl. Give them the right names.

A bytes constructor was using the wrong name for one of its fields.  It was
also assigning c_nil to the same field with type "qbytes_ptr_t". Fix the name
and change the nil assignment to use QBYTES_PTR_NULL.

[Reviewed by mppf]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20917 3a8e244f-b0f2-452b-bcba-4c88e055c3ca