Removing 2008 archive of HPC Challenge entries
This was another time capsule of HPCC entries, developed before we had
moved them into the release, I believe.  The one thing here that was
of value and may not appear anywhere else were some .txt files that I
used to use as a gimmick for categorizing lines of code into code
vs. comments vs. declarations vs. whitespace for the purposes of
sorting SLOC into different categories.  Still, the release versions
have changed enough that it might be just as fast to recreate this
information (or develop a script to try and determine it
automatically).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20922 3a8e244f-b0f2-452b-bcba-4c88e055c3ca