Factor chplrt.h [reviewed by gbt]
Motivation: The compilation commands for .c files in the runtime complain if
chplrt.h is not mentioned (exactly once!).  Given that every source file *must*
include it directly, its contents should be minimal up to the original intent.

Intent: The intent of chplrt.h was to inject macro definitions which control the
interaction of the runtime and generated C code with the environment, to ensure
portability, correctness and to provide some debugging support.  
These macros include redefinitions of malloc() and free(), redirecting them to internal
Chapel routines chpl_mem_alloc() and chpl_mem_free(), which enables the Chapel
memory debugging support.
Also redirected is printf().  (This may be obsolete.)
Finally, macros defining common types (in sys_basic.h) are provided.

Description: 
 - Removed from chplrt.h all (global) variable declarations, including those
   marked possibly obsolete by mppf.
 - Replaced #include "chpltypes.h" in chplrt.h with #include "sys_basic.h".
   chpltypes.h is intended for consumption by the generated code, so contains
   definitions which are not required by general runtime code.  Included in
   specific .c or .h files as needed: chpltimers.h, chplmemtrack.h.
 - Moved timer typedefs and function declarations out of chpltypes.h and into
   chpltimers.h.
 - In general, I like to see user includes precede system includes.  This coding
   style can result in shorter lists of #includes in an implementation.  So in
   some files that I touched for other reasons, I performed this transformation.
 - Moved comm diagnostic function declarations from chplrt.h to chpl-comm.h.
 - Moved chpl_compileCommand and other global declarations from chplrt.h to
   chplcgfns.h.  Noted in the header comment that chplcgfns also contains
   compiler-generated type declarations (not just function declarations).
 - Removed PRINTF_DEF and SIMPLE_TEST macros.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20923 3a8e244f-b0f2-452b-bcba-4c88e055c3ca