Fixing a bug relating to writing random streams
Prior to this commit, writing random streams would result in
a deadlock due to the read of the sync var it contains that
the default class writer causes.  Here, I'm explicitly defining
a writeThis() method to avoid this situation and print out the
user-visible constructor arguments.  Whether or not this is
the right/best thing to write out for a RandomStream is open
to debate, but it's better than a deadlock.

This also raises questions about whether reading/writing a sync
var should be a normal read/write of the variable (which is
what we've traditionally assumed) or whether it should display
some internal state about the sync var instead, or simply not
affect the full/empty state.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20924 3a8e244f-b0f2-452b-bcba-4c88e055c3ca