Fixing the compiler's handling of environment variable equivalents for the options of the types n and N. Clarifying related pieces of --help and the man page.
[reviewed by Tom]

* Changed the compiler behavior when looking at the environment
variable equivalent of an options of the type n or N. Depending
on the first character of the env var's string:

  Y y T t 1 - same as passing the option without --no,
  N n F f 0 - same as passing the option with --no,
  anything else - generates an error message explaining the above.

Cf. before it always resulted in the same setting regardless of the
environment variable contents. For an 'n' option it would set
it to "yes"; for 'N' it would to "no" (or visa versa?).

* Described the above behavior in the man page (chpl.txt).

* While in the man page, created a section on "ENVIRONMENT VARIABLES
FOR OPTIONS" that talks specifically about environment variable equivalents.
It precedes the existing ENVIRONMENT section that talk about the
build configuration-related and similar variables.

* Also changed the help text for the developer options -t/--html and
--html-user to indicate that they are toggle options.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20932 3a8e244f-b0f2-452b-bcba-4c88e055c3ca