Improve error messages for Array = scalar type mismatch errors
This patch is designed to produce a better error message for assignments
from scalars to arrays that result in a type mismatch.  For example:

        var A: [1..n] real(64);
        var r32: real(32);

        A = r32;

Prior to this patch, this case was falling into the final 'else' case of
proc =(a: [], b) in ChapelArray.chpl which was written assuming that you've
fallen into that case because you tried to do a parallel zip() iteration
over a and b but were not able to because b doesn't support parallel
iteration (whereas in this case, you've actually fallen into it because b
doesn't assign to a[i]).

My fix isn't completely satisfactory, but is a step in the right direction.
What I've done is to put a case at the beginning of
the nested conditional that asks

        (1) is b's type the element type of a?
        (2) is b a primitive type?

If either of these are true, then it implements the assignment using:

        forall aa in a do
          aa = b;

The effect is that:

(1) you get a simpler loop than zippered iteration in the case that they
    are the same type (which may have a bigger effect once we optimize
    standalone forall loops);

(2) that cases in which a primitive is being assigned that "just coerces"
    are handled similarly;

(3) that you get an error message when there is a type mismatch.  For
    example:

        testit.chpl:5: error: type mismatch in assignment from real(64) to
        real(32)

Moreover, the conditional is static, so should get folded out at compile
time and not result in runtime overhead (as with the other branches).

The reason I call this not completely satisfactory is that:

(a) it's unfortunate that it only fires when 'b' is a primitive type; it'd
really be nicer to say something like "if b is coercible to a's element
type." but currently, all that logic is in the compiler, not the modules
(we've discussed wanting to move this out into the language, but haven't
done it yet -- thinking about this en route to work this morning, it seems
very tractable to change to me, so I may make that my next pet project).

(b) Of course, if we had that test, then we'd still need a way to catch the
error case and say something useful about it.  At this moment I don't have
any good ideas about what characteristic would cause us to fall into that
case.  If 'b' is a singleton/not iteratable?

So, a step in an improved direction, though more work remains (as always?)



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20934 3a8e244f-b0f2-452b-bcba-4c88e055c3ca