Add support for multiple locale objects (sublocales) per node [reviewed by gbt]
This change is in preparation for hierarchical locale support in Chapel.  It
adds the capability to store and retrieve a sublocale ID in task-private data.
This capability is implemented in all known tasking layers.

On the Chapel side, the locale ID is composed of a node ID portion and a
sublocale ID portion.  Each of these is initially (arbrarily) allocated 32
bits.  A wide address now consists of this 64-bit localeID struct joined with a
64-bit (node-local, shared-memory) pointer.  

On the runtime side, the locale ID can be obtained by joining the node ID
obtained from the executing image (chpl_localeID) with the sublocale ID obtained
from task-private data (chpl_task_getSubLoc()).

Detailed changes:
- Replace integer locale ID with a struct containing node and sublocale IDs.
- Add an interface to the tasking layer, to get and set the task-private sublocale ID.
- Add a temporary struct declaration "c_locale_t" which mimics the declaration
  known to Chapel code and called chpl_localeID_t.
  It should be possible to remove this temporary declaration following a small
  refactoring of the runtime includes such that chplcgfns.h is *not* included in
  the build of the generated C code.  This remains as a work item.
- Pick off .node and .subloc fields as needed in runtime and generated code.
- Add new primitives to support localeID creation and field extraction.
- Rename "locale" as "node" globally.
- The locale ID is now passed as the first parameter to an ON block.
- Pipe through sublocale ID to backend task layer (not currently used).




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20938 3a8e244f-b0f2-452b-bcba-4c88e055c3ca