Fix for some gasnet testing failures due to r20938.
Gasnet tests were failing because I removed an optimization in build.cpp as part
of the dtLocale patch.

The optimization collapses chpl_int_to_locale(x).id with x.  I stubbed it out
because "id" in all of the existing Chapel code means "node id" whereas I expect
chpl_int_to_locale() to take a complete localeID as its argument.  Performing
the optimization as it was written meant we would be replacing an "int" result
with a "chpl_localeID_t" result.  I expected lots of code to break as a
consequence of the type mismatch.

A simple workaround is to recognize the same pattern, but return just the node
ID portion of the chpl_localeID_t argument.

A more complete solution involves replacing instances of "x.locale.id" in
existing source code with __primitive("_wide_get_node", x) wherever it occurs.
The ".locale.id" idiom and its corresponding optimization should be phased out.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20952 3a8e244f-b0f2-452b-bcba-4c88e055c3ca