Updated this future from a bug to a feature request on Michael's request
From the updated .future:

As originally filed, this test was working as advertised due to 

(a) the relative precedence of unary - and casting (casting wins, so
-1:int(8) means -(1:int(8)))

combined with

(b) math operators such as unary minus being supported only for 32-
and 64-bit ints (on the basis of "that's what most languages do" and
"that's what processors tend to support").


More recently, we've been talking about supporting math on all int
sizes, retaining their native size, so I've converted this into a
feature request for that capability on Michael's request.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20955 3a8e244f-b0f2-452b-bcba-4c88e055c3ca