Filing a bug against generic nested records
This is from one of my students.  It appears that nested generic
classes work and that non-nested generic records work but that nested
generic records do not.  It looks like it could be an easy fix for
someone familiar with constructor code (<cough>Tom<cough>) but there's
no rush to do so for my student's sake -- the assignment doesn't
require it and I believe the workarounds will be sufficient.

I also added a .bad file in anticipation of maybe having this feature
someday!  :)  Also the two working cases to lock in that behavior and
for comparison's sake.

From the .future:

When a generic class contains a generic inner record, resolution of a
constructor to that record fails.  Changing the inner record to a
class, or moving the record outside of the class works.  The error
message suggests that the cause of failure is due to a mismatch in
supplying 'outer' to the constructor call.  It's not clear to me
whether the fix would be to add the 'outer' argument to the
constructor's formals or to remove it from the actual's argument list.

Here's the error message:

innerRecord.chpl:5: In function 'chpl__initCopy':
innerRecord.chpl:5: error: unresolved call 'B(T=type int(64), v=int(64), c=int(64), outer=A(int(64)))'
innerRecord.chpl:5: note: candidates are: B(_mt: _MT, outer: A, type T, v: T, c)




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20965 3a8e244f-b0f2-452b-bcba-4c88e055c3ca