Fix bswap for GCC 4.7
This patch re-does bswap.h, which provides htobe64 and
friends using nonstandard GLIBC support, if available.
These macros are used in binary I/O. It also adds lots of
comments and a C language test case (so it's easy to
trouble-shoot or test different platforms).

The specific problem with the older bswap.h on GCC 4.7 is
that __bswap_64 is a function, not a macro. This patch
solves that problem by using bswap_64, from
byteswap.h (in glibc), which has always been a macro.

Thanks to Chris Wailes for reporting the problem with
GCC 4.7 and testing a fix. Thanks to Matt Baker for testing
further portability. And thanks to everyone for patience
with the testing noise and portability problems from this
feature.

Tested on:
linux gcc 4.4.6
linux clang 3.2
cray gcc 4.3.4
cray PGI 12.10
cray Intel cc 13.0.1
a cray CC
linux gcc 4.7.2
Mac OS X Snow Leopard with gcc and clang
Mac OS X Lion with gcc and clang
A Mac with gcc 4.7.2
Power7 with gcc 4.4.5

[reviewed by Tom]



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20966 3a8e244f-b0f2-452b-bcba-4c88e055c3ca