Clarified test/memory/shannon/memstatEquals.future .
(Consulted with Brad.)

This future seems to have started out when memstat-like options were
built into the compiler. Now that they are implemented at the Chapel
level, some of the bugs/semantic dudas have been worked out, but one
remains, which is now documented in .future.

Note that this future currently is not specific to memory-related options.

From the .future:
---
Given a bool configuration var/const, say 'bcv', we would like

  this command-line option...    to set this value for 'bcv':
    --bcv                           true
    --bcv=true                      true
    --bcv=false                     false
    --bcv=                          report an error
    --bcv[some other punctuation]   report an error

Currently such an error is not reported; --bcv= is treated as --bcv=true.
---

While there, clarified the message that the .chpl program prints out,
and clarified the desired error message in .good.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20967 3a8e244f-b0f2-452b-bcba-4c88e055c3ca