Bug fixes for QIO: EOF problem and Intel NANs
This patch fixes two problems.

First, Tom found an assert(err=EEOF), which was clearly
wrong, so he fixed it but that test started failing.

The problem with the EOF was that we were combining preemptive
mmaping (on file open) with channels using read/write (ie that
don't seek first). So, we add a check so that the preemptive
mmap is only used if we were planning to seek while reading next
(ie, the mode is PREAD/PWRITE or MMAP).

Second, test/io/ferguson/ctests were failing on Intel compilers.

The problem we ran into with the Intel compiler and the C tests
is that if we have double num = NAN, num = -num does not change
the sign bit. Also, 0.0*(1.0/0.0) produced 0.0 instead of NAN.
So, we use NAN and copysign to set a NAN positive.

[reviewed by Tom]



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20977 3a8e244f-b0f2-452b-bcba-4c88e055c3ca