Work around a file creation problem on certain test systems.
On certain of our test systems we have observed a problem which causes new
executable files not to be loaded into memory properly when they are run, even
though they have reasonable contents when read.  Perhaps this results from
some difference in how the kernel processes a new file's contents with mmap(2)
as opposed to with read(2), though I'm certainly not a kernel person and don't
know for sure.  For Chapel programs on Cray systems, this affects launcher
binaries.  It turns out that adding a "-p" option (preserve mode, owner, and
timestamps) to the cp(1) command that creates the executable file solves the
problem.  Since doing that has no negative consequences, I'm adding it as a
workaround for now.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20979 3a8e244f-b0f2-452b-bcba-4c88e055c3ca