Add RE2 to third-party
[reviewed by Tom]

The file hg_diff_g.patch is how we plan to patch RE2, along with adding the
files file_strings.h and file_strings.cc.  These changes allow RE2 to search a
FILE* or anything else that provides routines for getting the next character.
For Chapel, we will use this extra functionality in RE2 in order to make readf
with regular expressions work.

At some future point, we hope to include a patched version of RE2 in the Chapel
repository, but for now the script update-re2.sh will be run automatically
during the build process if RE2 is configured.  That script will apply our
patches to a downloaded RE2.
 


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20982 3a8e244f-b0f2-452b-bcba-4c88e055c3ca