Fix for failure of optimizeOnClauses due to r20938. [reviewed by gbt]
The fault was due to the failure to include the new primitive PRIM_GET_SUBLOC_ID
and PRIM_SET_SUBLOC_ID in the switch statement in isFastPrimitive() so that
'true' is returned for those cases.  Since those two primitives are now included
in every on_fn, on clause optimization was effectively disabled.

The solution involved making sure that there was an entry in said switch
statement for every known primitive.  This is enforced by an INT_FATAL being
called if the default label is reached.  Some other assertions were added at the
tail of that switch statement: Namely, GC primitives are flagged as
not-yet-implemented, GPU primitives are flagged as obsolete, and primitives
assumed to have been removed from the tree by earlier passes are flagged as
invalid.

As a result of this tedium, it was discovered that there were no entries in the
switch for PRIM_CHPL_COMM_GET, ...PUT, ...GET_STRD and ...PUT_STRD.  I added
these, and had those four cases return 'false': they are not fast
primitives because they are assumed to always entail communication.  Note that
in the prior implementation, any primitive not listed in the switch would have
been caught in the default case and 'false' returned.  So explicitly returning
false for those four cases is not a deviation from the former behavior.

I noticed that I had not supplied a definition for PRIM_WIDE_GET_SUBLOC in code
generation, so I added that.  I corrected an error in the implementation of code
generation for the PRIM_WIDE_GET_LOCALE and PRIM_WIDE_GET_NODE primitives: In
the case that the source was a wide class reference, the generated code was
fetching the wide pointer and then casting it to the desired type (localeID,
nodeID, etc.)  Now the desired field is selected from the wide pointer.  

I also noticed that there were trailing blanks in primitive.h, which I removed.
And while I was at it, I adjusted the indentation of some comments.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20987 3a8e244f-b0f2-452b-bcba-4c88e055c3ca