Revert tasking layers to being ignorant of the locale/node distinction.
[Reviewed by Tom.]

The r20973 commit changed chpl_task_addToTaskList() in a few tasking
layers, renaming the parameter identifying the task list's locale so as
to indicate that it identified a node instead of a locale.  After
discussion with the author of r20973, we've decided that because tasking
layers should not make assumptions about where they're running in a
locale hierarchy, this parameter should identify a locale after all,
rather than a node.  So, this commit renames it back how it was before.

While here, I also changed the name of the nanos tasking layer's file-static
local copy of the current locale so that it doesn't shadow the global one in
the runtime, changed a parameter name in the nanos_chpl_task_init() function
prototype in chpl_nanos.h to match the definition in chpl_main.cpp, and also
removed some spurious indentation from the prototype.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20990 3a8e244f-b0f2-452b-bcba-4c88e055c3ca