Remove the nanos tasking layer.
[Reviewed by Sung.]

This removes the nanos tasking layer and support thereof in the scripts
and documentation.  It has languished recently and we were working with
a pretty old copy already, so it seems unlikely to revive.  If interest
returns we can resurrect it, but for now retiring it will reduce the
maintenance burden.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20994 3a8e244f-b0f2-452b-bcba-4c88e055c3ca